requirements.txt
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
up-bank-api==1.1.0
